requirements.txt
fastapi
pydantic
uvicorn[standard]
requests
aiohttp
python-dotenv 
semantic-kernel[azure] 
aiosqlite

langgraph
langgraph-checkpoint-sqlite
langsmith
langchain-openai
langchain-tavily